Fixing type tests from merge
